overlay: test durability of change tracking snapshot
This is an overlay snapshot test that does NOT run with FSTYP=overlay
because FSTYP=overlay does not support flakey

Overlayfs snapshot assumed that copy up involved fsync of new file which
implies "metadata barrier", so the parent directory must exist in snapshot
after crash.  That is not true for a change tracking snapshot which only
records the parent directory, so that means that modifications to file may
end up not being tracked in snapshot after a crash.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>