src/t_dir_offset2: Check for inconsistent d_ino/st_ino
After unlink of a directory entry, that entry may still apear in getdents
results of an already open directory fd, but it should return a d_ino
value that is consistent with the already observed st_ino of that entry.

This is a regression test for kernel commit ....
("ovl: skip stale entries in merge dir cache iteration")

Remove redundant break condition from gendents read loop.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>