xfs: annotate fix commits for LTS backports from v5.17..v5.19
In preparation for backporting xfs fixes to stable kernels,
annotate some of the tests that pass after applying the backports.

Some of the annotated tests had the fix commit documented in header
comment, but the entire group of new fallocate tests did not refer
to the fix commit ("xfs: fallocate() should call file_modified()").

Signed-off-by: Amir Goldstein <amir73il@gmail.com>