open_by_handle: add a test for connectable file handles
This is a variant of generic/477 with connectable file handles.
This test uses load and store of file handles from a temp file to test
decoding connectable file handles after cycle mount and after renames.
Decoding connectable file handles after being moved to a new parent
is expected to fail.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>